better TreePopulator that works in build
For some reason the previous TreePopulator created trees that were not
correctly lit when the game is built. They looked fine in Unity editor,
but not from .exe when builded. Problem was the trees were created from
the script. Now I made it match the BearPopulator - not creating trees
from scratch in the script, but instead duplicating an original tree and
then deleting the orignial